fix(story): fix auth issue for the first time (#26)
## What?

This PR fixes the auth issue where the first-time authenticating user was not correctly redirected, but got in the infinite loop of redirection. After that, the user had to refresh the page, and it worked. That's why this bug wasn't discovered until now.

## Why?

JIRA: EXT-2202

Once a user is authenticated and redirected back to `/api/connect/callback?...`, at that point, unfortunately the cookie `sb.auth` is not set. I couldn't find out why, but it was the same with the previous `custom-app-examples` starter. Then why did it work there, and not here? When the authentication was done and the cookie was not found, the user was redirected again to `/api/connect`, and immediately the `app-extension-auth` redirects user back to `/api/connect/callback?` and this time there is a cookie set. So, this PR removes the cookie check and let it run the auth flow one more time. Now it works.

<!--
  Explain the **motivation** for making this change.
  What existing problem does the pull request solve?
  Are there any linked issues?
-->

## How to test? (optional)

<!--
  Demonstrate the code is solid.
  Example: The exact commands you ran and their output,
  screenshots / videos if the pull request changes UI.
-->